Reworks docs to use Spinx and Rst
Reworked the documentation to use Sphinx and reStructuredText markup
rather than mark down.

 * Added a contributing page on how to get this to build locally (you're
   gonna need python) as well as a style guideline
 * Converted the shouldBe and shouldNotBe documentation to reST
 * Added a simple example of bringing in source from an external file in
   the shouldBe documentation.

The example file pulls right from the unit test. I originally was going to
create a seperate project for just documentation purposes, but the
existing unit tests are so close that I didn't want to go down that route
without further input. Seems like some slight tweaking to the existing
unit tests on the formatting (and to throw in some Simpsons references)
and they could be used for the docs too.

Ultimately I think a "documentation" project will be needed. Little things
like referencing which line numbers to bring in would make working with
the unit test project and adjusting the using statements or adding a
comment could throw off the documentation.