Object Graph Comparison - ShouldBeEquivalentTo(...) (#411)
* Define public API.

* Add type comparison logic, message generation, and unit test.

* Add ValueType comparison logic and unit tests.

* Add null value comparison logic and unit tests.

* Add referential comparison logic and unit test.

* Add string comparison logic and unit tests.

* Add IEnumerable comparison logic and unit tests.
This change involves change the datatype of the 'path' variable from IList<T> to IEnumerable<T> to get LINQ functionality (and maybe some better performance due to fewer passes through the lists).

* Add child property comparison logic and unit tests.
This introduces the actual recursion into the comparison logic.
Some refactoring of code is included to call attention to the places where the object graph path is modified.

* Add infinite loop detection logic and unit tests.
This introduces a dictionary that keeps track of which objects have previously been (or are currently being) compared. Any new child comparisons will first check this dictionary; if the comparison has already been done between the current objects, then they are simply skipped. This should avoid infinite loop problems.
The dictionary is keyed by the actual object being compared, and the values represent a list of objects that the key has been compared against.